requirements.txt
flask>=3.0
gunicorn>=21.2
c2pa-python>=0.29
requests>=2.31
